+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -3,11 +3,16 @@
 Release History
 ===============
 
+2.1.10
+++++++
+* `policy assignment create`: support uri based parameters file
+* `policy set-definition update`: support uri based parameters and definitions files
+* `policy definition update`: fix handling of parameters and rules files
 * `resource show/update/delete/tag/invoke-action`: Fix issue where cross-subscription IDs did not properly honor the subscription ID.
 
 2.1.9
 +++++
-`group deployment create`: support uri based parameters file 
+* `group deployment create`: support uri based parameters file 
 * `policy assignment create/list/show`: support managed identity on policy assignments
 
 2.1.8